pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-bulletin-board"
description = "A simple bulletin board for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "bulletinboard",
    "eveonline",
]
license = {file = "LICENSE"}
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.8"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.0.0b1",
    "allianceauth-app-utils>=1.19.1",
    "django-ckeditor-5>=0.2.11",
    "unidecode",
]
[project.optional-dependencies]
tests-allianceauth-latest = [
    "coverage",
    "django-webtest",
    "Faker",
]
[project.urls]
Changelog = "https://github.com/ppfeufer/aa-bulletin-board/blob/master/CHANGELOG.md"
Documentation = "https://github.com/ppfeufer/aa-bulletin-board/blob/master/README.md"
Donations = "https://ko-fi.com/ppfeufer"
Homepage = "https://github.com/ppfeufer/aa-bulletin-board"
Source = "https://github.com/ppfeufer/aa-bulletin-board.git"
Tracker = "https://github.com/ppfeufer/aa-bulletin-board/issues"

[tool.hatch.build]
include = [
    "/aa_bulletin_board",
]

[tool.hatch.version]
path = "aa_bulletin_board/__init__.py"
